Finalized MoveToConfiguration
DONE:
- Tested MoveToConfiguration on the robot w/ lovelace
- Pulled the logic that will be shared with MoveToPose into a parent behavior.
- Implemented a generic MovetoConfiguration tree consisting of only the single MovetoConfiguration behavior
- Implemented MoveAbovePlate, MoveToRestingPosition, MoveToStowLocation as instances of the above tree
- Tested all three trees, both with specific ROS action calls and with the web app.
- Tested that cancellation works and immediately stops robot motion.
