+++ ada_feeding/ada_feeding/behaviors/__init__.py
@@ -1,5 +1,6 @@
 """
 This package contains custom py_tree behaviors for the Ada Feeding project.
 """
+from .move_to import MoveTo
 from .move_to_configuration import MoveToConfiguration
 from .move_to_dummy import MoveToDummy
