+++ ada_feeding/ada_feeding/trees/__init__.py
@@ -3,5 +3,6 @@ This package contains custom behavior trees that are used in the Ada Feeding
 project. Many of these trees implement the ActionServerBT interface, in order
 to be wrapped in a ROS action server. 
 """
-from .move_above_plate_tree import MoveAbovePlateTree
+from .move_to_tree import MoveToTree
+from .move_to_configuration_tree import MoveToConfigurationTree
 from .move_to_dummy_tree import MoveToDummyTree
