+++ ada_feeding/ada_feeding/trees/move_to_tree.py
@@ -1,34 +1,37 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-This module defines the MoveAbovePlate behavior tree and provides functions to
+This module defines the abstract MoveTo behavior tree and provides functions to
 wrap that behavior tree in a ROS2 action server.
+
+This module is intended to consolidate all the logic for generating feedback
+and results, which is shared across all MoveTo actions, into one place.
+Subclasses should only need to define create_move_to_tree.
 """
 
 # Standard imports
+from abc import ABC, abstractmethod
 import logging
-from typing import List
 
 # Third-party imports
 import py_trees
 from rclpy.node import Node
 
 # Local imports
-from ada_feeding.behaviors import MoveToConfiguration
 from ada_feeding.helpers import import_from_string
 from ada_feeding import ActionServerBT
 
 
-class MoveAbovePlateTree(ActionServerBT):
+class MoveToTree(ActionServerBT, ABC):
     """
-    Move the robot arm to a fixed above-plate position.
+    An abstract behvaior tree for any behavior that moves the robot and provides
+    results and feedback as specified in ada_feeding_msgs.action.MoveTo or
+    ada_feeding_msgs.action.AcquireFood.
     """
 
     def __init__(
         self,
         action_type_class: str,
-        joint_names: List[str],
-        joint_positions: List[float],
     ) -> None:
         """
         Initializes tree-specific parameters.
@@ -39,19 +42,9 @@ class MoveAbovePlateTree(ActionServerBT):
             e.g., "ada_feeding_msgs.action.MoveTo". The input of this action
             type can be anything, but the Feedback and Result must at a minimum
             include the fields of ada_feeding_msgs.action.MoveTo
-        joint_names: The names of the joints that the robot arm is moving.
-        joint_positions: The joint positions to move the robot arm to.
         """
         # Import the action type
         self.action_type_class = import_from_string(action_type_class)
-
-        # Store the joint names/positions
-        self.joint_names = joint_names
-        assert len(self.joint_names) == 6, "Must provide 6 joint names"
-        self.joint_positions = joint_positions
-        assert len(self.joint_positions) == len(
-            self.joint_names
-        ), "Must provide the same number of joint positions as joint names"
 
     def create_tree(
         self,
@@ -60,11 +53,7 @@ class MoveAbovePlateTree(ActionServerBT):
         node: Node,
     ) -> py_trees.trees.BehaviourTree:
         """
-        Creates the MoveAbovePlate behavior tree.
-
-        Currently, this only has one behavior in it, MoveToDummy. Eventually,
-        this should be replaced with a behavior tree that actually moves the
-        robot arm above the plate.
+        Create the behavior tree.
 
         Parameters
         ----------
@@ -77,23 +66,21 @@ class MoveAbovePlateTree(ActionServerBT):
         -------
         tree: The behavior tree that moves the robot above the plate.
         """
+        self.create_blackboard(name)
+        return self.create_move_to_tree(name, logger, node)
+
+    def create_blackboard(self, name: str) -> None:
+        """
+        Creates the blackboard for the behavior tree, and defines the blackboard
+        keys necessary to set the goal and send feedback.
+
+        Parameters
+        ----------
+        name: The name of the behavior tree.
+        """
         # Create the blackboard
         self.blackboard = py_trees.blackboard.Client(
             name=name + " Tree", namespace=name
-        )
-        # Inputs for MoveToConfiguration
-        self.blackboard.register_key(
-            key="joint_positions", access=py_trees.common.Access.WRITE
-        )
-        self.blackboard.register_key(
-            key="joint_names", access=py_trees.common.Access.WRITE
-        )
-        self.blackboard.register_key(
-            key="tolerance", access=py_trees.common.Access.WRITE
-        )
-        self.blackboard.register_key(key="weight", access=py_trees.common.Access.WRITE)
-        self.blackboard.register_key(
-            key="cartesian", access=py_trees.common.Access.WRITE
         )
         # Goal that is passed from the ROS2 Action Server
         self.blackboard.register_key(key="goal", access=py_trees.common.Access.WRITE)
@@ -114,16 +101,32 @@ class MoveAbovePlateTree(ActionServerBT):
             key="motion_curr_distance", access=py_trees.common.Access.READ
         )
 
-        # Write the inputs to MoveToConfiguration to blackboard
-        self.blackboard.joint_positions = self.joint_positions
-        self.blackboard.joint_names = self.joint_names
+    @abstractmethod
+    def create_move_to_tree(
+        self,
+        name: str,
+        logger: logging.Logger,
+        node: Node,
+    ) -> py_trees.trees.BehaviourTree:
+        """
+        Create the behavior tree. By the time this function is called, the
+        blackboard has already been created and been defined with the keys
+        necessary to set the goal and send feedback. Therefore, this function
+        is just responsible for defining the blackboard keys specific to the
+        movement behavior, and creating the tree.
 
-        # Create the tree
-        root = MoveToConfiguration(name, node)
-        root.logger = logger
-        tree = py_trees.trees.BehaviourTree(root)
+        Parameters
+        ----------
+        name: The name of the behavior tree.
+        logger: The logger to use for the behavior tree.
+        node: The ROS2 node that this tree is associated with. Necessary for
+            behaviors within the tree connect to ROS topics/services/actions.
 
-        return tree
+        Returns
+        -------
+        tree: The behavior tree that moves the robot above the plate.
+        """
+        raise NotImplementedError("create_move_to_tree not implemented")
 
     def send_goal(self, tree: py_trees.trees.BehaviourTree, goal: object) -> bool:
         """
