+++ ada_feeding/config/ada_feeding_action_servers.yaml
@@ -17,27 +17,19 @@ ada_feeding_action_servers:
     # Parameters for the MoveAbovePlate action
     MoveAbovePlate: # required
       action_type: ada_feeding_msgs.action.MoveTo # required
-      tree_class: ada_feeding.trees.MoveAbovePlateTree # required
+      tree_class: ada_feeding.trees.MoveToConfigurationTree # required
       tree_kws: # optional
         - action_type_class
-        - joint_names
         - joint_positions
       tree_kwargs: # optional
         action_type_class: ada_feeding_msgs.action.MoveTo # required
-        joint_names: # required
-          - j2n6s200_joint_1
-          - j2n6s200_joint_2
-          - j2n6s200_joint_3
-          - j2n6s200_joint_4
-          - j2n6s200_joint_5
-          - j2n6s200_joint_6
         joint_positions: # required
-          - -2.11666
-          - 3.34967
-          - 2.04129
-          - -2.30031
-          - -2.34026
-          - 2.9545
+          - -2.11666 # j2n6s200_joint_1
+          -  3.34967 # j2n6s200_joint_2
+          -  2.04129 # j2n6s200_joint_3
+          - -2.30031 # j2n6s200_joint_4
+          - -2.34026 # j2n6s200_joint_5
+          -  2.95450 # j2n6s200_joint_6
       tick_rate: 10 # Hz, optional, default: 30
 
     # Parameters for the AcquireFood action
@@ -57,15 +49,19 @@ ada_feeding_action_servers:
    # Parameters for the MoveToRestingPosition action
     MoveToRestingPosition: # required
       action_type: ada_feeding_msgs.action.MoveTo # required
-      tree_class: ada_feeding.trees.MoveToDummyTree # required
+      tree_class: ada_feeding.trees.MoveToConfigurationTree # required
       tree_kws: # optional
         - action_type_class
-        - dummy_plan_time
-        - dummy_motion_time
+        - joint_positions
       tree_kwargs: # optional
         action_type_class: ada_feeding_msgs.action.MoveTo # required
-        dummy_plan_time: 1.0 # secs
-        dummy_motion_time: 10.0 # secs
+        joint_positions: # required
+          - -1.94672 # j2n6s200_joint_1
+          -  2.51268 # j2n6s200_joint_2
+          -  0.35653 # j2n6s200_joint_3
+          - -4.76501 # j2n6s200_joint_4
+          -  5.99991 # j2n6s200_joint_5
+          -  4.99555 # j2n6s200_joint_6
       tick_rate: 10 # Hz, optional, default: 30
 
     # Parameters for the MoveToMouth action
@@ -83,15 +79,21 @@ ada_feeding_action_servers:
       tick_rate: 10 # Hz, optional, default: 30
 
     # Parameters for the MoveToStowLocation action
+    # TODO: We may be able to get a stow location that actually
+    # tucks to the side of the user's wheelchair.
     MoveToStowLocation: # required
       action_type: ada_feeding_msgs.action.MoveTo # required
-      tree_class: ada_feeding.trees.MoveToDummyTree # required
+      tree_class: ada_feeding.trees.MoveToConfigurationTree # required
       tree_kws: # optional
         - action_type_class
-        - dummy_plan_time
-        - dummy_motion_time
+        - joint_positions
       tree_kwargs: # optional
         action_type_class: ada_feeding_msgs.action.MoveTo # required
-        dummy_plan_time: 1.0 # secs
-        dummy_motion_time: 10.0 # secs
+        joint_positions: # required
+          - -1.52101 # j2n6s200_joint_1
+          -  2.60098 # j2n6s200_joint_2
+          -  0.32811 # j2n6s200_joint_3
+          - -4.00012 # j2n6s200_joint_4
+          -  0.22831 # j2n6s200_joint_5
+          -  3.87886 # j2n6s200_joint_6
       tick_rate: 10 # Hz, optional, default: 30
